change codedeploy script

+++ scripts/after-install.sh
@@ -1,4 +1,14 @@
-  #!/bin/bash
-  cd /var/frontend/
+#!/bin/bash
+cd /var/frontend/
+  
+# Check if the node_modules directory exists
+if [ -d "node_modules/" ]; then
+  # If it exists, remove it
+  echo "node_modules/ directory exists. Removing it..."
   rm -r node_modules/
-  pnpm install
+else
+  # If it doesn't exist, print a message
+  echo "node_modules/ directory does not exist. Nothing to do."
+fi 
+
+pnpm install
